force float casting to fix edge cases

+++ src/microns_to_nwb/tools/ophys/ophys.py
@@ -113,12 +113,12 @@ def add_functional_coregistration_to_plane_segmentation(
             cave_ids.append([np.nan])
 
         else:
-            pt_supervoxel_ids.extend(df["pt_supervoxel_id"].drop_duplicates().tolist())
-            pt_root_ids.extend(df["pt_root_id"].drop_duplicates().tolist())
-            pt_x_positions.extend(df["pt_position_x"].drop_duplicates().tolist())
-            pt_y_positions.extend(df["pt_position_y"].drop_duplicates().tolist())
-            pt_z_positions.extend(df["pt_position_z"].drop_duplicates().tolist())
-            cave_ids.append(df["id"].values.tolist())
+            pt_supervoxel_ids.extend(df["pt_supervoxel_id"].drop_duplicates().astype(np.float64).tolist())
+            pt_root_ids.extend(df["pt_root_id"].drop_duplicates().astype(np.float64).tolist())
+            pt_x_positions.extend(df["pt_position_x"].drop_duplicates().astype(np.float64).tolist())
+            pt_y_positions.extend(df["pt_position_y"].drop_duplicates().astype(np.float64).tolist())
+            pt_z_positions.extend(df["pt_position_z"].drop_duplicates().astype(np.float64).tolist())
+            cave_ids.append(df["id"].astype(np.float64).values.tolist())
 
     plane_segmentation.add_column(
         name="cave_ids",
